+++ tests/test_tcod.py
@@ -170,6 +170,22 @@ def test_noise_errors():
     with pytest.raises(ValueError):
         noise.sample_ogrid(np.ogrid[:2,:2,:2])
 
+@pytest.mark.parametrize('implementation',
+    [tcod.noise.SIMPLE, tcod.noise.FBM, tcod.noise.TURBULENCE])
+def test_noise_pickle(implementation):
+    rand = tcod.random.Random(tcod.random.MERSENNE_TWISTER, 42)
+    noise = tcod.noise.Noise(2, implementation, rand=rand)
+    noise2 = copy.copy(noise)
+    assert (noise.sample_ogrid(np.ogrid[:3,:1]) ==
+            noise2.sample_ogrid(np.ogrid[:3,:1])).all()
+
+def test_noise_copy():
+    rand = tcod.random.Random(tcod.random.MERSENNE_TWISTER, 42)
+    noise = tcod.noise.Noise(2, rand=rand)
+    noise2 = pickle.loads(pickle.dumps(noise))
+    assert (noise.sample_ogrid(np.ogrid[:3,:1]) ==
+            noise2.sample_ogrid(np.ogrid[:3,:1])).all()
+
 def test_color_class():
     assert tcod.black == tcod.black
     assert tcod.black == (0, 0, 0)